♻️ refactor(server): isClientDisconnect → expect/actual (native-safe)
